adjustments

+++ library-OwnProject/src/main/java/com/binary/library/repositories/LibraryRepository.java
@@ -5,13 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LibraryRepository extends CrudRepository<Library, Integer> {
+public interface LibraryRepository extends CrudRepository<Library, Long> {
 
-    //count() --> returns the number of entities
-    //findALL() --> returns all items of given type
-    //findById() --> returns one item by id
-    //delete(T entity) --> delete an entity
-    //deleteAll() - deletes all the entities in the repository
-    //save (T entity) - saves an entity
-    //saveAll(); -- saves all the entities
+
 }
